Update index.md
I suggest to shorten the signing up text and rank the text in order or priority. Most important is that the newby volunteer understands the step by step process. 

The DIVD slack community is nice to have and a side step
The DIVD introduction course is relevant to prepare for the interview, and not earlier, assuming the orientation has been done prior to the sign up step.